moved scripts, changed Invoke function with method name to a direct method call
